+++ backend/src/main/java/ca/metricalsky/winston/entity/SettingsEntity.java
@@ -1,0 +1,30 @@
+package ca.metricalsky.winston.entity;
+
+import jakarta.persistence.Basic;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "settings")
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class SettingsEntity {
+
+    @Id
+    @Column(name = "setting_name")
+    private String name;
+
+    @Basic
+    @Column(name = "setting_value")
+    private String value;
+}