+++ backend/src/main/java/ca/metricalsky/winston/repository/AuthorRepository.java
@@ -1,7 +1,9 @@
 package ca.metricalsky.winston.repository;
 
 import ca.metricalsky.winston.entity.AuthorEntity;
+import ca.metricalsky.winston.entity.view.AuthorChannelView;
 import ca.metricalsky.winston.entity.view.AuthorDetailsView;
+import ca.metricalsky.winston.entity.view.AuthorVideoView;
 import ca.metricalsky.winston.entity.view.VideoStatisticsView;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -55,4 +57,37 @@ public interface AuthorRepository extends JpaRepository<AuthorEntity, String> {
             GROUP BY v.id
             """)
     List<VideoStatisticsView> findVideoStatisticsByAuthorId(String id);
+
+    @Query("""
+            SELECT
+                ch.title AS channelTitle,
+                ch.customUrl AS channelHandle,
+                COUNT(DISTINCT v.id) AS videoCount,
+                COUNT(co.id) AS totalCommentCount,
+                COUNT(co.parentId) AS replyCount,
+                MIN(co.publishedAt) AS firstCommentedAt,
+                MAX(co.publishedAt) AS lastCommentedAt
+            FROM ChannelEntity ch
+                JOIN VideoEntity v ON ch.id = v.channelId
+                JOIN CommentEntity co ON v.id = co.videoId
+            WHERE co.author.displayName = :displayName
+            GROUP BY ch.id
+            """)
+    List<AuthorChannelView> findAuthorChannelsByDisplayName(String displayName);
+
+    @Query("""
+            SELECT
+                v.id AS videoId,
+                v.title AS videoTitle,
+                v.thumbnailUrl AS videoThumbnailUrl,
+                COUNT(co.id) AS totalCommentCount,
+                COUNT(co.parentId) AS replyCount,
+                MIN(co.publishedAt) AS firstCommentedAt,
+                MAX(co.publishedAt) AS lastCommentedAt
+            FROM VideoEntity v
+                JOIN CommentEntity co ON v.id = co.videoId
+            WHERE co.author.displayName = :displayName
+            GROUP BY v.id
+            """)
+    List<AuthorVideoView> findAuthorVideosByDisplayName(String displayName);
 }